Feature/rp 1132

+++ src/components/organisms/Tables/Team/index.tsx
@@ -15,6 +15,8 @@ import {
 
 import api from 'api'
 import { routes } from 'routes'
+import { percentCalculate } from 'components/utils/percentCalculate'
+import { toast } from '@stardust-ds/react'
 
 import { GRID_TEMPLATE, HEADERS } from '../../Forms/Project/constants'
 import { Shelf } from './Shelf'
@@ -43,21 +45,31 @@ export const Team = () => {
 
   function handleUpdateUser(user_id: number) {
     const newTeam = Team.map((item) =>
-      item.user_id === user_id ? { ...item } : item
+      item.user_id === user_id ? { ...item, } : item
     )
     api.put(routes.project.userProjects(Number(user_id)), newTeam)
 
     setValue('team', newTeam)
+    return newTeam
+
+
   }
 
   function removeUser(user_id: number) {
     if (project_id) {
-      api.delete(routes.project.userProjects(Number(project_id)), {
-        data: { user_id }
-      })
+      api.delete(routes.project.userProjects(Number(project_id)), { data: { user_id } })
+
     }
-    const newTeam = Team.filter((item) => item.user_id !== user_id)
+    const newTeam = Team.filter(
+      (item) => item.user_id !== user_id
+    )
+
     setValue('team', newTeam)
+    toast({
+      title: 'Profissional removido com sucesso',
+      type: 'success',
+    })
+    
   }
 
   const Table = useMemo(() => {
@@ -68,21 +80,25 @@ export const Team = () => {
         </LoadingWrapper>
       )
 
-    return Team.map((props) => (
-      <Shelf
-        key={props.user_id}
-        config={{
-          template: GRID_TEMPLATE,
-          options: POPOVER_OPTIONS(
-            props.user_id,
-            props.is_active,
-            props.name
-          )
-        }}
-        {...{ props }}
-      />
-    ))
+    return (
+      <>
+        {Team.map((props) => (
+          < Shelf
+            key={props.user_id}
+            config={{
+              template: GRID_TEMPLATE,
+              options: POPOVER_OPTIONS(props.user_id, props.is_active, props.name)
+            }
+            }
+            {...{ props }}
+          />
+        ))}
+
+      </>
+    )
   }, [isLoading, Team])
+
+
 
   return (
     <Main>
